stratum, stratum-proxy: Fix new_shares_sum in handle_submit_result
- Change new_shares_sum to u64 as defined in the spec
- new_shares_sum in handle_submit_result is equal to v2_target

+++ protocols/stratum/src/v2/messages.rs
@@ -150,7 +150,7 @@ pub struct SubmitSharesSuccess {
     pub channel_id: u32,
     pub last_seq_num: u32,
     pub new_submits_accepted_count: u32,
-    pub new_shares_sum: u32,
+    pub new_shares_sum: u64,
 }
 
 #[id(0x1du8)]
